refactor : define Compose Presentation Structure
