+++ composeApp/src/commonMain/kotlin/org/patchnote/patchnote/presentation/model/UserModel.kt
@@ -1,0 +1,4 @@
+package org.patchnote.patchnote.presentation.model
+
+class UserModel {
+}
