+++ composeApp/src/commonMain/kotlin/org/patchnote/patchnote/presentation/screen/App.kt
@@ -1,4 +1,4 @@
-package org.patchnote.patchnote
+package org.patchnote.patchnote.presentation.screen
 
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.foundation.Image
@@ -15,6 +15,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import org.jetbrains.compose.resources.painterResource
 import org.jetbrains.compose.ui.tooling.preview.Preview
+import org.patchnote.patchnote.Greeting
 
 import patchnote11.composeapp.generated.resources.Res
 import patchnote11.composeapp.generated.resources.compose_multiplatform
