+++ composeApp/src/commonMain/kotlin/org/patchnote/patchnote/presentation/state/UserContract.kt
@@ -1,0 +1,4 @@
+package org.patchnote.patchnote.presentation.state
+
+class UserContract {
+}
