+++ composeApp/src/commonMain/kotlin/org/patchnote/patchnote/presentation/viewmodel/UserViewModel.kt
@@ -1,0 +1,4 @@
+package org.patchnote.patchnote.presentation.viewmodel
+
+class UserViewModel {
+}
